Actualización de producto-microservicio (Funcional)

+++ common-ecommerce-microservicio/src/main/java/com/itsel/common/dto/ProductoDTO.java
@@ -5,66 +5,63 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
 public class ProductoDTO {
-	
-	private Long id;
-	
-	@NotNull(message = "El nombre no puede ser nulo")
-	@NotBlank(message = "El nombre no puede estar en blanco")
-	private  String nombre;
-	
-	@NotNull(message = "La descripcion no puede ser nulo")
-	@NotBlank(message = "La descripcion no puede estar en blanco")
-	private  String descripcion;
-	
-	@NotNull(message = "El precio no puede ser nulo")
+    
+    private Long id;
+    
+    @NotNull(message = "El nombre no puede ser nulo")
+    @NotBlank(message = "El nombre no puede estar en blanco")
+    private String nombre;
+    
+    @NotNull(message = "La descripcion no puede ser nulo")
+    @NotBlank(message = "La descripcion no puede estar en blanco")
+    private String descripcion;
+    
+    @NotNull(message = "El precio no puede ser nulo")
     @PositiveOrZero(message = "El precio no puede ser negativo")
-	private  Long precio;
-	
-	@NotNull(message = "El stock no puede ser nulo")
+    private Double precio;  // Cambiado de Long a Double
+    
+    @NotNull(message = "El stock no puede ser nulo")
     @PositiveOrZero(message = "El stock no puede ser negativo")
-	private  Long stock;
+    private Integer stock;  // Cambiado de Long a Integer
 
-	public Long getId() {
-		return id;
-	}
+    // Getters y Setters
+    public Long getId() {
+        return id;
+    }
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	public String getNombre() {
-		return nombre;
-	}
+    public String getNombre() {
+        return nombre;
+    }
 
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
 
-	public String getDescripcion() {
-		return descripcion;
-	}
+    public String getDescripcion() {
+        return descripcion;
+    }
 
-	public void setDescripcion(String descripcion) {
-		this.descripcion = descripcion;
-	}
+    public void setDescripcion(String descripcion) {
+        this.descripcion = descripcion;
+    }
 
-	public Long getPrecio() {
-		return precio;
-	}
+    public Double getPrecio() {
+        return precio;
+    }
 
-	public void setPrecio(Long precio) {
-		this.precio = precio;
-	}
+    public void setPrecio(Double precio) {
+        this.precio = precio;
+    }
 
-	public Long getStock() {
-		return stock;
-	}
+    public Integer getStock() {
+        return stock;
+    }
 
-	public void setStock(Long stock) {
-		this.stock = stock;
-	}
-	
-	
-
-
+    public void setStock(Integer stock) {
+        this.stock = stock;
+    }
 }
